implementar  cambios para seguir el  contrato de openapi

+++ Application/Interfaces/IOrder/IOrderCommand.cs
@@ -7,7 +7,7 @@ namespace Application.Interfaces.IOrder
     public interface IOrderCommand
     {
         Task<Order> CreateOrder(Order order);
-        Task<Order> UpdateStatusItemOrder(int orderId, int itemId, OrderItemUpdateRequest request);
-        Task<Order> UpdateOrder(ICollection<OrderItem> request, int id, decimal newPrice);
+        Task<Order> UpdateStatusItemOrder(long orderId, long itemId, OrderItemUpdateRequest request);
+        Task<Order> UpdateOrder(ICollection<OrderItem> request, long id, decimal newPrice);
     }
 }
